Strip the syntax of operator type instead of injecting it
We could inject, but then we'd have to also inject all the symbols
we're comparing against (and use the "compare" procedure to do so).
Instead, because these identifiers are treated as plain symbols, it's
easier to strip the syntax.

This was always broken, but actually got exposed due to CHICKEN core
commit 41d48cdba2e46645c7dd157525652aa3953aed6c, which fixed an
obscure bug related to double renaming.